backend/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Redis
redis==5.0.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 配置管理
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP 客户端
httpx==0.26.0

# 工具
python-dotenv==1.0.0

# 测试
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0
